split things up into types, lib & exe
