+++ Hackage/Mirror/Types.hs
@@ -1,0 +1,24 @@
+module Hackage.Mirror.Types where
+
+import qualified Codec.Archive.Tar as Tar
+import           Data.ByteString (ByteString)
+import qualified Data.ByteString.Lazy as BL
+
+data Options = Options
+    { verbose     :: Bool
+    , rebuild     :: Bool
+    , mirrorFrom  :: String
+    , mirrorTo    :: String
+    , s3AccessKey :: String
+    , s3SecretKey :: String
+    }
+
+data Package = Package
+    { packageName       :: !String
+    , packageVersion    :: !String
+    , packageCabal      :: !BL.ByteString
+    , packageIdentifier :: !ByteString
+    , packageTarEntry   :: !Tar.Entry
+    }
+
+data PathKind = UrlPath | S3Path | FilePath
